merge the publicsuffix crate and the psl crate
This commit merges both crates, which used to live in different repos.
The `psl` crate is now meant to supersede the `publicsuffix` crate.
I have redesigned and rewritten both crates to address
https://github.com/rushmorem/publicsuffix/issues/15.

The `psl` crate is the main crate. It's now very lean and lightweight.
It is the lowest level in the suite of crates. See the README for
how to use it.